move rollup to shared module
